chore: incomplete refactor
