+++ lib/external/array-deque.ts
