+++ lib/index.ts
@@ -8,5 +8,7 @@ import * as math from './utils/math';
 import * as SV from './utils/string-view';
 import * as strings from './utils/strings';
 import * as grid from './utils/grid';
+import * as stringset from './external/string-set';
+import * as arraydeque from './external/array-deque';
 
-export { array, bench, bits, color, data, general, math, strings, SV, grid };
+export { array, bench, bits, color, data, general, math, strings, SV, grid, stringset, arraydeque };
